api/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic==2.11.3
pydantic-settings==2.8.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
openai==1.71.0
httpx==0.28.1
python-multipart==0.0.20
